// src/app/models/participantTypes.model.ts
export class participantTypesModel {
  Id: number;
  Name: string;
  Description: string;
  CompanyId: number;
  CreatedById: number;
  UpdatedById: number;
  CreatedByRoleId: number;
  UpdatedByRoleId: number;
  CreatedDateTime: Date;
  UpdatedDdateTime: Date;
  ParameterCarrier: string;
}